Improvements on Rework for expansion
