+++ src/data/thargoidData.ts
@@ -1,0 +1,33 @@
+import { Thargoid } from './dataFormat';
+
+const cyclops: Thargoid = {
+    name: "Cyclops",
+    armourRating: 100,
+    regenPerSecond: 3.2,
+    exertHp: 160
+};
+
+const basilisk: Thargoid = {
+    name: "Basilisk",
+    armourRating: 140,
+    regenPerSecond: 7.2,
+    exertHp: 360
+};
+
+const medusa: Thargoid = {
+    name: "Medusa",
+    armourRating: 175,
+    regenPerSecond: 10,
+    exertHp: 500
+};
+
+const hydra: Thargoid = {
+    name: "Hydra",
+    armourRating: 230,
+    regenPerSecond: 12.8,
+    exertHp: 640
+};
+
+export const THARGOID_TYPES = [
+    cyclops, basilisk, medusa, hydra
+];
