+++ src/data/weaponData.ts
